Moved dict accessors into Resource
