+++ tracs/plugins/json.py
@@ -2,12 +2,10 @@ from __future__ import annotations
 
 from datetime import timedelta
 from logging import getLogger
-from typing import Any, Callable, Dict, List, Optional, Tuple, Union
+from typing import Any, Union
 
 from cattrs.preconf.orjson import make_converter
-from more_itertools import last, rstrip
 from orjson import dumps as save_json, loads as load_json, OPT_APPEND_NEWLINE, OPT_INDENT_2, OPT_SORT_KEYS
-
 from tracs.handlers import ResourceHandler
 from tracs.pluginmgr import importer, resourcetype
 from tracs.resources import ResourceType
@@ -32,30 +30,6 @@ class JSONHandler( ResourceHandler ):
 
 	def save_raw( self, data: Any, **kwargs ) -> bytes:
 		return save_json( data, option=JSONHandler.OPTIONS, default=serialize )
-
-	# additional helpers for json access
-
-	def _value( self, *args, parent: Dict = None, conv: Callable, default: Any = None ) -> Any:
-		parent, item = parent or self.raw, last( args )
-		for s in rstrip( args, lambda e: e is item ):
-			parent = parent[s]
-
-		try:
-			return conv( val ) if ( val := parent[item] ) not in [ '', None, 0, 0.0 ] else default
-		except (KeyError, TypeError, ValueError):
-			return default
-
-	def float( self, *args, parent: Dict = None, default=None ) -> Optional[float]:
-		return self._value( *args, parent=parent, conv=float, default=default )
-
-	def list( self, *args, parent: Dict = None, default=None ) -> Optional[List]:
-		return self._value( *args, parent=parent, conv=list, default=default )
-
-	def int( self, *args, parent: Dict = None, default=None ) -> Optional[int]:
-		return self._value( *args, parent=parent, conv=int, default=default )
-
-	def str( self, *args, parent: Dict = None, default=None ) -> Optional[str]:
-		return self._value( *args, parent=parent, conv=str, default=default )
 
 class DataclassFactoryHandler( JSONHandler ):
